[CORE24-254] task(list): fix tests, display actual data
