+++ apps/core-frontend/src/styles.scss
