+++ libs/prm-components/src/lib/components/EntityList/Cell.component.tsx
@@ -1,5 +1,6 @@
 import { Td } from '@chakra-ui/react';
 import { Entity } from '@nrcno/core-models';
+import _ from 'lodash';
 
 import { EntityUIConfig } from '../../config';
 
@@ -9,7 +10,12 @@ type Props = {
 };
 
 export const Cell: React.FC<Props> = ({ entity, field }) => {
-  const rawValue = field.path.reduce<any>((acc, key) => acc[key], entity);
-  const value = field.format ? field.format(rawValue) : rawValue;
+  const rawValue = _.get(entity, field.path.join('.'));
+
+  const value = !rawValue
+    ? ''
+    : field.format
+      ? field.format(rawValue)
+      : rawValue;
   return <Td width={`${field.width}rem`}>{value}</Td>;
 };
