+++ libs/prm-components/src/lib/pages/EntityListPage/EntityListPage.test.tsx
@@ -1,6 +1,7 @@
 import { EntityType } from '@nrcno/core-models';
 import { render } from '@testing-library/react';
 import axios from 'axios';
+import { createMemoryRouter, RouterProvider } from 'react-router-dom';
 
 import { PrmProvider } from '../../prm.context';
 
@@ -10,15 +11,22 @@ const axiosInstance = axios.create();
 
 describe('EntityListPage', () => {
   it('should render successfully', () => {
-    const { baseElement } = render(
-      <PrmProvider
-        axiosInstance={axiosInstance}
-        entityType={EntityType.Participant}
-        entityId={undefined}
-      >
-        <EntityListPage />
-      </PrmProvider>,
-    );
+    const router = createMemoryRouter([
+      {
+        path: '/',
+        element: (
+          <PrmProvider
+            axiosInstance={axiosInstance}
+            entityType={EntityType.Participant}
+            entityId={undefined}
+          >
+            <EntityListPage />
+          </PrmProvider>
+        ),
+      },
+    ]);
+    const { baseElement } = render(<RouterProvider router={router} />);
+
     expect(baseElement).toBeTruthy();
     expect(baseElement.getElementsByTagName('h2')[0].textContent).toBe(
       'participants',
