+++ package.json
@@ -26,6 +26,7 @@
     "framer-motion": "^11.1.7",
     "helmet": "^7.1.0",
     "knex": "^3.1.0",
+    "lodash": "^4.17.21",
     "nocache": "^4.0.0",
     "passport": "^0.7.0",
     "passport-openidconnect": "^0.1.2",
